requirements.txt
Flask>=2.2
flask-sqlalchemy>=3.0
flask-migrate>=4.0
python-dotenv>=1.0
flask-jwt-extended>=4.5
flask-cors>=4.0
Flask-Limiter>=3.5
marshmallow>=3.20
marshmallow-sqlalchemy>=0.28.0
oracledb>=1.3            # Oracle DB driver
SQLAlchemy>=2.0          # Required for oracledb dialect
bcrypt>=4.0
structlog>=23.1
boto3>=1.34              # optional S3 storage
pytest>=7.4
pytest-flask>=1.3
gunicorn>=21.2
